% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{apply_transform}
\alias{apply_transform}
\title{Apply the given transform to x,y coordds}
\usage{
apply_transform(transform, coords_df)
}
\arguments{
\item{transform}{matrix transform to apply}

\item{coords_df}{data.frame containing x,y coords}
}
\value{
data.frame with updated coordinates
}
\description{
Apply the given transform to x,y coordds
}
