% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-path-arc.R
\name{arc_to_df}
\alias{arc_to_df}
\title{Convert \code{<path>} arc to a data.frame of coordinates}
\usage{
arc_to_df(params, npoints)
}
\arguments{
\item{params}{named list of arc parameters from the path segment i.e. xo, y0,
x, y, arcflag, sweepflag, rx, ry, rot (degrees)}

\item{npoints}{number of points along arc}
}
\value{
data.frame of coodinates of arc
}
\description{
Reference \url{https://svgwg.org/svg2-draft/implnote.html#ArcConversionEndpointToCenter}
}
\details{
ToDo: actual rotation of arc coords by 'phi'
}
