% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-rect.R
\name{create_rect_df}
\alias{create_rect_df}
\title{Create a data.frame for a rectangle with rounded corners}
\usage{
create_rect_df(x, y, width, height, rx, ry, npoints)
}
\arguments{
\item{x, y, width, height, rx, ry}{rect parameters}

\item{npoints}{number of points along each of the rounded corners}
}
\value{
data.frame of (x, y) coordinates
}
\description{
Create a data.frame for a rectangle with rounded corners
}
