% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_calc_all_styles}
\alias{css_calc_all_styles}
\title{Create a named list to store computed styles indexed by the xpath of each element}
\usage{
css_calc_all_styles(svg, user_css = NULL)
}
\arguments{
\item{svg}{root node of svg}

\item{user_css}{user supplied CSS (as a single text string)}
}
\value{
list of style lookup by xpath
}
\description{
Create a named list to store computed styles indexed by the xpath of each element
}
