% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-units.R
\name{css_value_as_numeric}
\alias{css_value_as_numeric}
\title{Convert a string value from CSS into a numeric value}
\usage{
css_value_as_numeric(x, percentage_as_fraction = TRUE, ...)
}
\arguments{
\item{x}{Character string of a CSS value e.g. "12", "12px", "3em", "47\%"}

\item{percentage_as_fraction}{Should percentages like "12%" be returned
as a fraction (e.g. 0.12)?  Default: TRUE.   If FALSE then returns
the percentage as a numeric value (e.g. 12)}

\item{...}{other arguments passed to \code{css_length_as_pixels()}}
}
\value{
a numeric value as best we can with limited knowledge
}
\description{
Convert a string value from CSS into a numeric value
}
