% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame}
\alias{data_frame}
\title{data.frame creation}
\usage{
data_frame(...)
}
\arguments{
\item{...}{named arguments}
}
\value{
data.frame
}
\description{
20x Faster than 'data.frame()', but with *ZERO* sanity checking and
no support for row.names
}
