% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-ellipse.R
\name{ellipse_to_df}
\alias{ellipse_to_df}
\title{Convert Ellipse parameters into a data.frame of coords}
\usage{
ellipse_to_df(cx, cy, rx, ry, npoints)
}
\arguments{
\item{cx, cy, rx, ry}{ellipse parameters}

\item{npoints}{number of points around ellipse}
}
\value{
data.frame of x,y coordinates around the ellipse
}
\description{
\code{cx, cy, rx, ry} are attribute parameters of the \code{<ellipse>} tag.
\code{npoints} is defined by the ser.
}
