% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_element_by_id}
\alias{get_element_by_id}
\title{Get an element from an XML document by ID. Return NA if ID does not exist.}
\usage{
get_element_by_id(svg, id)
}
\arguments{
\item{svg}{xml document containing svg}

\item{id}{id, #id, url(#id), url('#id')}
}
\value{
SVG element with the given id, otherwise NA
}
\description{
This function is called when a \code{<use>} tag references another element
or a a colour tag references a gradient element.
}
