% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-style-to-gp-vp.R
\name{gpar_to_df}
\alias{gpar_to_df}
\title{Convert a gpar object to a data.frame}
\usage{
gpar_to_df(gp)
}
\arguments{
\item{gp}{gpar object with a single specification. no vectors allowed}
}
\value{
data.frame with single row.
}
\description{
This converts a gpar object to a 1-row data.frame representation. This is
used when accumulating a data.frame representation to return to the user.
}
\details{
Linear and Radial gradient objects are replaced with their first colour.
Although this could be something fancier e.g. list object, or some text
representation of the gradient parameters. Maybe just \code{deparse1()}?
}
