% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{Break a string into labelled tokens based upon a set of patterns}
\usage{
lex(text, regexes, verbose = FALSE)
}
\arguments{
\item{text}{a single character string}

\item{regexes}{a named vector of regex strings.  Each string represents
a regex to match a token, and the name of the string is the
label for the token.  Each regex can contain an explicit
captured group using the standard \code{()} brackets. If a regex
doesn't not define a captured group then the entire regex will
be captured. The regexes
will be processed in order such that an early match takes
 precedence over any later match.}

\item{verbose}{print more information about the matching process. default: FALSE}
}
\value{
a named character vector with the names representing the token type
        with the value being the element extracted by the corresponding
        regular expression.
}
\description{
Break a string into labelled tokens based upon a set of patterns
}
\examples{
\dontrun{
lex("hello there 123.45", regexes=c(number=re$number, word="(\\\\w+)", whitespace="(\\\\s+)"))
}

}
