% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertinyicons.R
\name{load_supertinyicon}
\alias{load_supertinyicon}
\title{Load an SVG icon from the \code{supertinyicons} icon pack}
\usage{
load_supertinyicon(
  name,
  obj_type = c("grob", "data.frame", "list", "debug", "svg"),
  ...
)
}
\arguments{
\item{name}{name of icon. e.g. \code{"twitter"}.  If the exact name is not
found, then the icon with
the closest matching name is returned.  See \code{supertinyicon_names}
for a full list of supported icons in this set.}

\item{obj_type}{What kind of R object to return - choices
\code{c('grob', 'data.frame', 'list', 'debug', 'svg')}. Default: 'grob'.
See documentation for \code{read_svg()} for more details.}

\item{...}{other arguments passed to \code{read_svg()}}
}
\description{
Load an SVG icon from the \code{supertinyicons} icon pack
}
