% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{parse_command_to_matrix}
\alias{parse_command_to_matrix}
\title{Parse tokens representing a single transform instruction into a matrix}
\usage{
parse_command_to_matrix(tokens)
}
\arguments{
\item{tokens}{character vector of c(command, number, number, ...)}
}
\value{
matrix representing this transform
}
\description{
Parse tokens representing a single transform instruction into a matrix
}
