% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-parse.R
\name{parse_svg_group}
\alias{parse_svg_group}
\title{Parse xml representing some SVG}
\usage{
parse_svg_group(elem, state)
}
\arguments{
\item{elem}{as SVG grouping element e.g. <svg> <g> <clipPath>}

\item{state}{environment containing state information for this SVG}
}
\description{
Parse xml representing some SVG
}
