% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-gradient.R, R/svg-parse.R
\name{parse_svg_radialGradient}
\alias{parse_svg_radialGradient}
\alias{parse_svg_linearGradient}
\alias{parse_svg_use}
\alias{parse_svg_path}
\alias{parse_svg_polyline}
\alias{parse_svg_polygon}
\alias{parse_svg_line}
\alias{parse_svg_rect}
\alias{parse_svg_circle}
\alias{parse_svg_ellipse}
\alias{parse_svg_text}
\title{Parse an SVG tag of given type}
\usage{
parse_svg_radialGradient(elem, state)

parse_svg_linearGradient(elem, state)

parse_svg_use(elem, state)

parse_svg_path(elem, state)

parse_svg_polyline(elem, state)

parse_svg_polygon(elem, state)

parse_svg_line(elem, state)

parse_svg_rect(elem, state)

parse_svg_circle(elem, state)

parse_svg_ellipse(elem, state)

parse_svg_text(elem, state)
}
\arguments{
\item{elem}{SVG element which is of known type}

\item{state}{environment containing state information for this SVG}
}
\value{
grob
}
\description{
Parse an SVG tag of given type
}
