% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-path.R
\name{parse_svg_path_d}
\alias{parse_svg_path_d}
\title{Parse path into a list of parameters and meta-information}
\usage{
parse_svg_path_d(svg_path)
}
\arguments{
\item{svg_path}{SVG path string (single character string)}
}
\value{
list of lists of information - 1 list per element in the path.
        each element is itself a list of parameters and meta-information
        about the path element e.g. index of the element in the path
}
\description{
Parse path into a list of parameters and meta-information
}
