% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{parse_transform_string_to_matrices}
\alias{parse_transform_string_to_matrices}
\title{Parse a transform string to a list of matrices - 1 per command. Mainly for debugging}
\usage{
parse_transform_string_to_matrices(transform_string)
}
\arguments{
\item{transform_string}{e.g. "rotate(10) translate(3, 3)"}
}
\value{
list of matrices
}
\description{
Note there can be multiple transform instructions in a transform string
}
