% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{parse_transform_string_to_matrix}
\alias{parse_transform_string_to_matrix}
\title{Parse a transform string to asingle transform matrix}
\usage{
parse_transform_string_to_matrix(transform_string)
}
\arguments{
\item{transform_string}{e.g. "rotate(10) translate(3, 3)"}
}
\value{
combined transform marix
}
\description{
Note there can be multiple transform instructions in a transform string
}
