% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-path.R
\name{path_list_expand_beziers}
\alias{path_list_expand_beziers}
\title{Prepare complete list of control point coordinates for all beziers}
\usage{
path_list_expand_beziers(path_list, state)
}
\arguments{
\item{path_list}{path as returned by parse_svg_path_d}

\item{state}{named list of current state. Default list(x=0, y=0)}
}
\value{
\code{path_list} with all bezier elements now having 'x' and 'y'
        vectors with coords of the control points.
}
\description{
Prepare complete list of control point coordinates for all beziers
}
