% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-path.R
\name{path_list_to_abs_path_list}
\alias{path_list_to_abs_path_list}
\title{Convert path relative coordinates to absolute coordinates}
\usage{
path_list_to_abs_path_list(path_list, state = list(x = 0, y = 0))
}
\arguments{
\item{path_list}{path as returned by \code{parse_svg_path_d()}}

\item{state}{named list of current state. Default \code{list(x=0, y=0)}}
}
\value{
\code{path_list} with only absolute elements
}
\description{
Convert path relative coordinates to absolute coordinates
}
