% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-path.R
\name{path_list_to_df}
\alias{path_list_to_df}
\title{Convert a path list to a data.frame}
\usage{
path_list_to_df(path_list, state = list(x = 0, y = 0))
}
\arguments{
\item{path_list}{path as returned by parse_svg_path_d}

\item{state}{named list of current state. Default list(x=0, y=0)}
}
\value{
data.frame of x,y coordinates
}
\description{
Convert a path list to a data.frame
}
