% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_dfs}
\alias{rbind_dfs}
\title{Internal function for perfroming do.call(rbind...)}
\usage{
rbind_dfs(dfs)
}
\arguments{
\item{dfs}{list of data.frames}
}
\value{
single data.frame
}
\description{
This function expands all data.frames to have the same columns, whereas
base R just complains a throws an error
}
