% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertinyicons.R
\name{read_file}
\alias{read_file}
\title{Brute force file reader that is not concerned with file endings}
\usage{
read_file(conn)
}
\arguments{
\item{conn}{connection. For svgparser this is always an \code{unz()} connection to
the zipped version of the icon set}
}
\value{
single string representing file contents
}
\description{
\code{readLines()} in combination with \code{unz()} fails o read the last
line of a file if it does not end in a CR/LF.  I haven't found a workaround
besides a manual read loop like below.
}
\details{
This function reads all character in a file regardless of line endings or
file endings.
}
