% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertinyicons.R
\name{shotgun_match}
\alias{shotgun_match}
\title{Pick the closest matching word from a list of words.}
\usage{
shotgun_match(word, words)
}
\arguments{
\item{word}{user word}

\item{words}{list of all words}
}
\description{
As long as 'word' & 'words' are not empty, this function should always
return a valid member of 'words'.
}
