% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-style-to-gp-vp.R
\name{style_to_gpar}
\alias{style_to_gpar}
\title{Convert a list of style information for an element to a \code{gpar} object}
\usage{
style_to_gpar(style, state)
}
\arguments{
\item{style}{named list of style properties for an element. E.g.
\code{list(stroke = 'black')}}

\item{state}{state information. See inline documentation for \code{read_svg} for
more information.}
}
\value{
gpar object
}
\description{
Convert a named list of CSS style information into a gpar object. E.g.
\code{list(stroke = 'black')} is converted to \code{gpar(col = 'black')}.
}
\details{
This function handles stroke and fill colours (and separate alphas for each),
fonts, linear and radial gradients, line parameters (width, join and end
types etc).

Not yet done:
\itemize{
\item{Conversion of line type information from CSS to gpar}
}
}
