% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-style-to-gp-vp.R
\name{style_to_viewport}
\alias{style_to_viewport}
\title{Convert a list of style information to a viewport object}
\usage{
style_to_viewport(style, transform, state)
}
\arguments{
\item{style}{named list of style properties}

\item{transform}{transform matrix of element being clipped}

\item{state}{state information}
}
\value{
viewport object
}
\description{
Currently this only supports setting the clipping path on the viewport.
}
\details{
Masks would also be set here, but I don't have any good examples to
debug with, so leaving this for now.
}
