% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertinyicons.R
\docType{data}
\name{supertinyicon_names}
\alias{supertinyicon_names}
\title{Names of icons in \code{supertinyicons} set}
\format{
An object of class \code{character} of length 297.
}
\usage{
supertinyicon_names
}
\description{
Names of icons in \code{supertinyicons} set
}
\keyword{datasets}
