% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-colour.R
\name{svg_colour_to_r_colour}
\alias{svg_colour_to_r_colour}
\title{Convert an SVG colour to an R colour}
\usage{
svg_colour_to_r_colour(col, state)
}
\arguments{
\item{col}{svg colour}

\item{state}{environment containing state information for this SVG}
}
\value{
valid R colour
}
\description{
Handles: False colouring, hex colours, oct colours, radial gradients.
}
\details{
More complex CSS colour specifications like \code{rgb(10, 20, 30, 0.5)}
are handled by \code{cssparser::css_colour_to_hex()}
}
