% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{update_transform}
\alias{update_transform}
\title{Update a transform matrix with an element}
\usage{
update_transform(transform, elem)
}
\arguments{
\item{transform}{matrix}

\item{elem}{svg element}
}
\value{
new transform matrix
}
\description{
Update a transform matrix with an element
}
