% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{update_transform_with_string}
\alias{update_transform_with_string}
\title{Update a transform matrix with the new SVG transform string}
\usage{
update_transform_with_string(transform, transform_string)
}
\arguments{
\item{transform}{matrix}

\item{transform_string}{the transform string which is a presentation
attribute on an element}
}
\value{
new transform matrix
}
\description{
Update a transform matrix with the new SVG transform string
}
